separate app and server
